moved validation to middleware + moved schema to schemas
