+++ routes/api/contacts.js
@@ -3,15 +3,17 @@ const express = require("express");
 const router = express.Router();
 
 const { ctrlContacts } = require("../../controllers");
+const { validateBody } = require("../../middlewares");
+const { contactSchema } = require("../../schemas");
 
 router.get("/", ctrlContacts.getAll);
 
 router.get("/:contactId", ctrlContacts.getContact);
 
-router.post("/", ctrlContacts.postContact);
+router.post("/", validateBody(contactSchema), ctrlContacts.postContact);
 
 router.delete("/:contactId", ctrlContacts.remContact);
 
-router.put("/:contactId", ctrlContacts.updContact);
+router.put("/:contactId", validateBody(contactSchema), ctrlContacts.updContact);
 
 module.exports = router;
