+++ schemas/index.js
@@ -1,0 +1,5 @@
+const contactSchema = require("./contacts.js");
+
+module.exports = {
+  contactSchema,
+};
